move linux file handling to separate class
